pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mc-to-synq"
version = "0.1.0"
description = "Migrate monitors from Monte Carlo to SYNQ (Coalesce Quality) with a single command."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ryan Schlenz" },
]
keywords = ["monte-carlo", "synq", "coalesce", "data-observability", "migration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
mc-to-synq = "mc_to_synq.cli:app"

[project.urls]
Homepage = "https://github.com/ryanschlenz/mc-to-synq"
Issues = "https://github.com/ryanschlenz/mc-to-synq/issues"

[tool.setuptools.packages.find]
include = ["mc_to_synq*"]

[tool.setuptools.package-data]
mc_to_synq = ["../config/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
